[#61] Fix error if item has non-existing groups
Even though an error is nice to immediately hint an developer that there
is something wrong with the setup, it's even more important to still
function. If code for printing items is reused in multiple places but
different groups/level of filtering is needed, an error would complicate
the whole situation for the developer.